update: rename from HelloWorld and make basic welcome screen
